Moves video related code to its own file
